refactor(harness): decompose loop_.rs into retry, history, and cwd modules
Splits the four distinct concerns in loop_.rs into focused sub-modules:
- retry.rs: max_retries, is_rate_limit, complete_with_retry
- history.rs: load_history, persist_user_message
- cwd.rs: resolve_session_cwd, resolve_session_cwd_with_root
- loop_.rs: run, run_tool_dispatch_loop only

Updates harness.spec.md to document the new module boundaries and invariants.

Closes #73
